requirements.txt
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
ffmpeg-python==0.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
future==1.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
waitress==3.0.2
Werkzeug==3.1.3
